Implement equality and hashing of partitions
This allows to quickly check partitions on equality. Further, it
allows to store partitions in a Python set. This allows to quickly
compute the number of different partitions in a large set of
partitions.

Note that modifying partitions that are stored in a set or even used
as keys in a dict leads to undefined behavior.