CKBDynamic: limit amount of edges touched during perturbation
Every time a community is perturbed, its amount of edges is
re-sampled. Previously, a community would always adjust to this
re-sampled amount of edges and then additionally perturb the given
amount of edges. Now the amount of edges inserted or removed due to
the adjustment is subtracted from the amount of edges perturbed.