CKBDynamic: link edge probabilities to curent community size
This means the edge probability now always depends on the current
community size and not for example in the case of a birth event, on
the final community size. This breaks the
CustomCommunitySizeDistribution, which most likely needs some custom
fitting of some function parameters.

This also includes a rewrite of parts of the phases of a merge event
to make them easier to understand. Events now happen faster by always
rounding up the number of nodes that shall be added or removed in a
time step instead of rounding them down.

Event probabilities are now adjusted after every single community
event and directly take into account the number of community
memberships after the completion of the event. This avoids an
over-compensation of a surplus or lack of memberships that might lead
to the inverse problem.