QTM: fix an edge case where the choice might not be optimal
This fixes an edge case where the node to move is a leaf that has
exactly one sibling and the parent path does not only contain
neighbors of the node to move.

To avoid complicated changes, this removes the property that no move
is performed if the number of edits is the same in the case sortPaths
is on. For determining if a move was executed, however, only the
number of edits is considered so the algorithm still stops if no
improvement was possible in a whole round. If randomness is enabled,
the criteria is instead if either an improvement was possible or if
there was more than one choice.

These changes also remove the reliance on neighbors in other methods
in preparation for other changes. Further, the case that there are no
close enough neighbors is not a special case anymore, which fixes
issues with the number of choices in rootData not being correctly set.