QTM: if sortPaths and randomNess is on, ignore duplicate options
This changes the code to never adopt children unless the parent is at
the lower end of a simple path. The reason is that when adopting the
then only child, the resulting graph is identical to the graph
obtained from choosing the lowermost node of the path as parent and
adopting all its children. This ensures that we choose uniformly at
random among all possible sets of edits of minimum size incident to
the moved node.