QTM: fix the debug code not finding equally good positions
With sortPaths on, the current parent and current children are no
longer available and in also general, it seems much safer to compute
the minimum number of edits from scratch instead of relying on the old
number (which is no longer verified). This also adds some additional
debug output that helps manual debugging.